+++ src/collective/classification/tree/__init__.py
@@ -4,3 +4,4 @@ from zope.i18nmessageid import MessageFactory
 
 
 _ = MessageFactory("collective.classification.tree")
+SOURCE_RESULTS_LEN = 20